Move schemas outside of app
