--- apps/puzzlify/app/schemas/adventure.ts
@@ -1,8 +1,0 @@
-import { z } from "zod";
-
-export const adventureSchema = z.object({
-  name: z.string().min(1),
-  startDate: z.string().datetime(),
-});
-
-export type Adventure = z.infer<typeof adventureSchema>;
