--- apps/puzzlify/app/schemas/riddle.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-export const riddleSubmission = z.object({
-  answer: z.string().min(1).toLowerCase(),
-});
-
-export type RiddleSubmission = z.infer<typeof riddleSubmission>;
