--- apps/puzzlify/app/schemas/user.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-export const userSchema = z.object({
-  email: z.string().email(),
-  // TODO: Add password requirements
-  password: z.string().min(1),
-});
