+++ apps/puzzlify/components/HintForm.tsx
@@ -2,7 +2,7 @@
 
 import { addHint } from "@/actions/host/create-stage-hint";
 import { editHint } from "@/actions/host/edit-stage-hint";
-import { hintSchema } from "@/app/schemas/stage";
+import { hintSchema } from "@/schemas/stage";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useHookFormAction } from "@next-safe-action/adapter-react-hook-form/hooks";
 import { Adventure, Hint, Stage } from "@prisma/client";
