+++ apps/puzzlify/components/RegistrationForm.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { userSchema } from "@/app/schemas/user";
+import { userSchema } from "@/schemas/user";
 import { createUser } from "@/actions/user/create-user";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useHookFormAction } from "@next-safe-action/adapter-react-hook-form/hooks";
